Organize SignalProtocol extensions into file
