+++ Playground.playground/Pages/Exploration.xcplaygroundpage/Contents.swift
@@ -4,3 +4,14 @@ import ReactiveKit
 import PlaygroundSupport
 
 //: Explore ReactiveKit here
+
+enum MyError: Error {
+    case unknown
+}
+
+let a = Signal<Int, Error>(sequence: 0...4, interval: 0.5)
+let b = SafeSignal(sequence: 0...2, interval: 2)
+
+a.amb(with: b.castError()).observeNext { (e) in
+    print(e)
+}
