+++ Playground.playground/contents.xcplayground
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<playground version='6.0' target-platform='ios' display-mode='raw'>
+<playground version='6.0' target-platform='ios' display-mode='raw' last-migration='1020'>
     <pages>
         <page name='Creating Signals'/>
         <page name='Exploration'/>
