+++ ReactiveKit/SignalProtocol+ErrorHandling.swift
@@ -1,0 +1,211 @@
+//
+//  The MIT License (MIT)
+//
+//  Copyright (c) 2016-2019 Srdan Rasic (@srdanrasic)
+//
+//  Permission is hereby granted, free of charge, to any person obtaining a copy
+//  of this software and associated documentation files (the "Software"), to deal
+//  in the Software without restriction, including without limitation the rights
+//  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+//  copies of the Software, and to permit persons to whom the Software is
+//  furnished to do so, subject to the following conditions:
+//
+//  The above copyright notice and this permission notice shall be included in
+//  all copies or substantial portions of the Software.
+//
+//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+//  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+//  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+//  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+//  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+//  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+//  THE SOFTWARE.
+//
+
+import Foundation
+
+extension SignalProtocol {
+
+    /// Transform error by applying `transform` on it.
+    public func mapError<F>(_ transform: @escaping (Error) -> F) -> Signal<Element, F> {
+        return Signal { observer in
+            return self.observe { event in
+                switch event {
+                case .next(let element):
+                    observer.next(element)
+                case .failed(let error):
+                    observer.failed(transform(error))
+                case .completed:
+                    observer.completed()
+                }
+            }
+        }
+    }
+
+    /// Branches out error into another signal.
+    public func branchOutError() -> (Signal<Element, Never>, Signal<Error, Never>) {
+        let shared = shareReplay()
+        return (shared.suppressError(logging: false), shared.toErrorSignal())
+    }
+
+    /// Branches out mapped error into another signal.
+    public func branchOutError<F>(_ mapError: @escaping (Error) -> F) -> (Signal<Element, Never>, Signal<F, Never>) {
+        let shared = shareReplay()
+        return (shared.suppressError(logging: false), shared.toErrorSignal().map(mapError))
+    }
+
+    /// Converts signal into non-failable signal by suppressing the error.
+    public func suppressError(logging: Bool, file: String = #file, line: Int = #line) -> Signal<Element, Never> {
+        return Signal { observer in
+            return self.observe { event in
+                switch event {
+                case .next(let element):
+                    observer.next(element)
+                case .failed(let error):
+                    observer.completed()
+                    if logging {
+                        print("Signal at \(file):\(line) encountered an error: \(error)")
+                    }
+                case .completed:
+                    observer.completed()
+                }
+            }
+        }
+    }
+
+    /// Converts signal into non-failable signal by feeding suppressed error into a subject.
+    public func suppressAndFeedError<S: SubjectProtocol>(into listener: S, logging: Bool = true, file: String = #file, line: Int = #line) -> Signal<Element, Never> where S.Element == Error {
+        return feedError(into: listener).suppressError(logging: logging, file: file, line: line)
+    }
+
+    /// Recovers the signal by propagating default element if error happens.
+    public func recover(with element: Element) -> Signal<Element, Never> {
+        return Signal { observer in
+            return self.observe { event in
+                switch event {
+                case .next(let element):
+                    observer.next(element)
+                case .failed:
+                    observer.next(element)
+                    observer.completed()
+                case .completed:
+                    observer.completed()
+                }
+            }
+        }
+    }
+
+    /// Restart the operation in case of failure at most `times` number of times.
+    public func retry(times: Int) -> Signal<Element, Error> {
+        guard times > 0 else { return toSignal() }
+        return Signal { observer in
+            var remainingAttempts = times
+            let serialDisposable = SerialDisposable(otherDisposable: nil)
+            var attempt: (() -> Void)?
+            attempt = {
+                serialDisposable.otherDisposable?.dispose()
+                serialDisposable.otherDisposable = self.observe { event in
+                    switch event {
+                    case .next(let element):
+                        observer.next(element)
+                    case .failed(let error):
+                        if remainingAttempts > 0 {
+                            remainingAttempts -= 1
+                            attempt?()
+                        } else {
+                            attempt = nil
+                            observer.failed(error)
+                        }
+                    case .completed:
+                        attempt = nil
+                        observer.completed()
+                    }
+                }
+            }
+            attempt?()
+            return BlockDisposable {
+                serialDisposable.dispose()
+                attempt = nil
+            }
+        }
+    }
+
+    /// Retries the failed signal when other signal produces an element.
+    /// - parameter other: Signal that triggers a retry attempt.
+    /// - parameter shouldRetry: Retries only if this returns true for a given error. Defaults to always returning true.
+    public func retry<S: SignalProtocol>(when other: S, if shouldRetry: @escaping (Error) -> Bool = { _ in true }) -> Signal<Element, Error> where S.Error == Never {
+        return Signal { observer in
+            let serialDisposable = SerialDisposable(otherDisposable: nil)
+            var attempt: (() -> Void)?
+            attempt = {
+                serialDisposable.otherDisposable?.dispose()
+                let compositeDisposable = CompositeDisposable()
+                serialDisposable.otherDisposable = compositeDisposable
+                compositeDisposable += self.observe { event in
+                    switch event {
+                    case .next(let element):
+                        observer.next(element)
+                    case .completed:
+                        attempt = nil
+                        observer.completed()
+                    case .failed(let error):
+                        if shouldRetry(error) {
+                            compositeDisposable += other.observe { otherEvent in
+                                switch otherEvent {
+                                case .next:
+                                    attempt?()
+                                case .completed:
+                                    attempt = nil
+                                    observer.failed(error)
+                                }
+                            }
+                        } else {
+                            observer.failed(error)
+                        }
+                    }
+                }
+            }
+            attempt?()
+            return serialDisposable
+        }
+    }
+
+    /// Error-out if `interval` time passes with no emitted elements.
+    public func timeout(after interval: Double, with error: Error, on queue: DispatchQueue = DispatchQueue(label: "reactive_kit.timeout")) -> Signal<Element, Error> {
+        return Signal { observer in
+            var completed = false
+            let timeoutWhenPossible: () -> Disposable = {
+                return queue.disposableAfter(when: interval) {
+                    if !completed {
+                        completed = true
+                        observer.failed(error)
+                    }
+                }
+            }
+            var lastSubscription = timeoutWhenPossible()
+            return self.observe { event in
+                lastSubscription.dispose()
+                observer.on(event)
+                completed = event.isTerminal
+                lastSubscription = timeoutWhenPossible()
+            }
+        }
+    }
+
+    /// Maps failable signal into a non-failable signal of errors. Ignores `.next` events.
+    public func toErrorSignal() -> Signal<Error, Never> {
+        return Signal { observer in
+            return self.observe { taskEvent in
+                switch taskEvent {
+                case .next:
+                    break
+                case .completed:
+                    observer.completed()
+                case .failed(let error):
+                    observer.next(error)
+                    observer.completed()
+                }
+            }
+        }
+    }
+}
