+++ ReactiveKit/SignalProtocol+Result.swift
@@ -1,25 +1,28 @@
 //
-//  SignalProtocol+Result.swift
-//  ReactiveKit-iOS
+//  The MIT License (MIT)
 //
-//  Created by Srdan Rasic on 24/03/2019.
-//  Copyright © 2019 DeclarativeHub. All rights reserved.
+//  Copyright (c) 2016-2019 Srdan Rasic (@srdanrasic)
+//
+//  Permission is hereby granted, free of charge, to any person obtaining a copy
+//  of this software and associated documentation files (the "Software"), to deal
+//  in the Software without restriction, including without limitation the rights
+//  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+//  copies of the Software, and to permit persons to whom the Software is
+//  furnished to do so, subject to the following conditions:
+//
+//  The above copyright notice and this permission notice shall be included in
+//  all copies or substantial portions of the Software.
+//
+//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+//  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+//  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+//  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+//  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+//  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+//  THE SOFTWARE.
 //
 
 import Foundation
-
-public protocol _ResultProtocol {
-    associatedtype Value
-    associatedtype Error: Swift.Error
-    var _unbox: Result<Value, Error> { get }
-}
-
-extension Result: _ResultProtocol {
-
-    public var _unbox: Result {
-        return self
-    }
-}
 
 extension SignalProtocol {
 
@@ -79,5 +82,18 @@ extension SignalProtocol where Element: _ResultProtocol, Error == Never {
     /// A failure result will trigger signal failure.
     public func getValues() -> Signal<Element.Value, Element.Error> {
         return (castError() as Signal<Element, Element.Error>).getValues()
+    }
+}
+
+public protocol _ResultProtocol {
+    associatedtype Value
+    associatedtype Error: Swift.Error
+    var _unbox: Result<Value, Error> { get }
+}
+
+extension Result: _ResultProtocol {
+
+    public var _unbox: Result {
+        return self
     }
 }
